Refactor to inject page object into steps via DI from Spring.
No functional changes.
